change feeding time duration to decimal

+++ database/migrations/2022_12_05_093650_create_feeding_times_table.php
@@ -17,7 +17,7 @@ class CreateFeedingTimesTable extends Migration
             $table->id();
             $table->time('time');
             $table->json('days');
-            $table->integer('duration');
+            $table->decimal('duration');
             // $table->string('state', 20);
             $table->boolean('isActive')->default(true);
             $table->timestamps();
